filter: Moved influxUnescape to influx.Unescape
Preparation for tag ordering & downsampler features.
