+++ influx/unescape.go
@@ -1,15 +1,29 @@
-package filter
+// Copyright 2018 Jump Trading
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package influx
 
 import "bytes"
 
-// influxUnescape returns a new slice containing the unescaped version
-// of in.
+// Unescape returns a new slice containing the unescaped version of
+// in.
 //
 // This is the same as Unescape() from
 // github.com/influxdata/influxdb/pkg/escape.
 // It's copied here because it's not worth vendoring all of influxdb
 // just for this.
-func influxUnescape(in []byte) []byte {
+func Unescape(in []byte) []byte {
 	if bytes.IndexByte(in, '\\') == -1 {
 		return in
 	}
@@ -47,7 +61,7 @@ func influxUnescape(in []byte) []byte {
 			}
 		}
 		out = append(out, in[i])
-		i += 1
+		i++
 	}
 	return out
 }
